v1.11.1: диапазонные оценки отзывов (4.5-5.0)
Поле «Оценка отзыва» теперь принимает либо фикс (`5`), либо диапазон
(`4.5-5.0`, `4,5-5,0` с запятой). Per-review random в [min, max] с шагом 0.1.

- Options::parseReviewsRatingRange / pickRandomRating helpers
- ReviewsGenerator пикает random per-scenario
- ApiClient передаёт float, CustomBackend пишет float в N-свойство
- BPRG (server) рендерит звёзды по round и хранит float в rating-поле JSON

Дефолт для новых установок — 4.5-5.0. Существующие с reviews_default_rating
мигрируют автоматически в фикс-оценку.

+++ lib/apiclient.php
@@ -110,7 +110,9 @@ class ApiClient
             'settings' => [
                 'min_words' => (int)($settings['min_words'] ?? 20),
                 'max_words' => (int)($settings['max_words'] ?? 60),
-                'rating' => (int)($settings['rating'] ?? 5),
+                // float, чтобы поддержать диапазонные оценки (4.5-5.0). Серверный
+                // BPRG принимает оба формата — int или float.
+                'rating' => (float)($settings['rating'] ?? 5),
                 'custom_prompt' => (string)($settings['custom_prompt'] ?? ''),
                 'temperature' => (float)($settings['temperature'] ?? 0.8),
                 'creative_mode' => !empty($settings['creative_mode']),
